composer, namespace, api moved from http to https
