+++ composer.json
@@ -1,0 +1,20 @@
+{
+  "name": "schmiddim/glosbe-translate",
+  "description": "forked from vhenriquez/Glosbe-Translate-PHP",
+  "homepage": "https://github.com/vhenriquez/Glosbe-Translate-PHP",
+  "license": "GPL-3.0",
+  "autoload": {
+    "psr-0": {"Translate\\": "./src"}
+  },
+  "description": "API Wrapper for Glosbe",
+  "authors": [
+    {
+      "name": "Michael Schmitt",
+      "email": "schmiddim@gmx.at"
+    }
+  ],
+  "minimum-stability": "dev",
+  "require": {
+
+  }
+}
